requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
idna==2.10
openpyxl==3.0.7
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
